src/core/config.py
from typing import Dict, Any
from pydantic import BaseModel
from functools import lru_cache
import os
from dotenv import load_dotenv

# Load .env file
load_dotenv()

class Settings(BaseModel):
    # API Settings
    API_HOST: str = os.getenv("API_HOST", "0.0.0.0")
    API_PORT: int = int(os.getenv("API_PORT", "8000"))
    DEBUG: bool = os.getenv("API_DEBUG", "True").lower() == "true"
    
    # OpenAI Settings
    OPENAI_API_KEY: str = os.getenv("OPENAI_API_KEY", "")
    OPENAI_MODEL: str = os.getenv("OPENAI_MODEL", "gpt-3.5-turbo")
    OPENAI_TEMPERATURE: float = float(os.getenv("OPENAI_TEMPERATURE", "0.7"))
    OPENAI_MAX_TOKENS: int = int(os.getenv("OPENAI_MAX_TOKENS", "800"))
    
    # Project Info
    PROJECT_NAME: str = "AI 智慧客服系統"
    VERSION: str = "1.0.0"
    DESCRIPTION: str = "基於 AI 的智慧客服系統"
    
    # Database Settings
    MONGO_URI: str = os.getenv("MONGO_URI", "mongodb://localhost:27017")
    MONGO_DB_NAME: str = os.getenv("MONGO_DB_NAME", "customer_service")
    REDIS_URI: str = os.getenv("REDIS_URI", "redis://localhost:6379")
    
    # Collections
    CHAT_COLLECTION: str = "chats"
    USER_COLLECTION: str = "users"
    KNOWLEDGE_COLLECTION: str = "knowledge"
    
    # Security Settings
    JWT_SECRET_KEY: str = os.getenv("JWT_SECRET_KEY", "your-secret-key-here")
    JWT_ALGORITHM: str = os.getenv("JWT_ALGORITHM", "HS256")
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
    
    # Vector Store Settings
    VECTOR_STORE_TYPE: str = os.getenv("VECTOR_STORE_TYPE", "chroma")
    EMBEDDING_MODEL: str = os.getenv("EMBEDDING_MODEL", "sentence-transformers/all-mpnet-base-v2")
    
    # LLM Settings
    LLM_CONTEXT_WINDOW: int = 4096
    LLM_MAX_MEMORY_SIZE: int = 5
    
    # Cache Settings
    CACHE_TTL: int = 3600  # in seconds

    class Config:
        env_file = ".env"

@lru_cache()
def get_settings() -> Settings:
    """Get cached settings instance"""
    return Settings()

# Create settings instance
settings = get_settings()