requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# AI/LLM
openai>=1.3.5
pydantic>=2.7.4
pydantic-settings>=2.0.3
transformers>=4.35.1
sentence-transformers>=2.2.2
torch>=2.1.1

# Database
motor>=3.3.2
redis[hiredis]>=5.0.1
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
chromadb>=0.4.15

# Utils
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
prometheus-client>=0.18.0
pytest>=7.4.3
requests>=2.31.0